backend/src/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.110.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.2
pydantic_core==2.16.3
pymongo==4.6.2
pyparsing==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
